Go: allow GET /api/v1/users to anyone via CORS
This pulls in github.com/go-chi/cors, a middleware to manage CORS in
go-chi, and uses it to allow GET to /api/v1/users (and only that
endpoint) to anyone.

This is slightly awkward, as it seems like that library wasn't really
made to set up handler-specific CORS settings, only global. As such, we
have to manually specify an r.Options handler, that is the cors
middleware wrapping a nil handler.

Tested manually by doing:

    $ curl -H "Origin: http://example.com" \
        -H "Access-Control-Request-Method: GET" \
        -X OPTIONS --verbose \
        127.0.0.1:3000/api/v1/users
    [...]
    Access-Control-Allow-Methods: GET
    Access-Control-Allow-Origin: *

I can implement some unit/integration tests if someone would point me
into a recommended way of doing that within the go-chi ecosystem.

Replaces #30.