Go: allow GET /api/v1/users to anyone via CORS (#31)
This pulls in github.com/go-chi/cors, a middleware to manage CORS in
go-chi, and uses it to allow GET to /api/v1/users (and only that
endpoint) to anyone.

This is slightly awkward, as it seems like that library wasn't really
made to set up handler-specific CORS settings, only global. As such, we
have to manually specify an r.Options handler, that is the cors
middleware wrapping a nil handler.

Replaces #30.