refactor: Use UserEntry for storage interfaces.
Creating models package was a mistake. To make single
packages more independent, I will have to fully replace
usage of models types with data structures oriented with
given approach.

This is the first step in this big process, that enables
usage of OnlineUsers storage interface and doesn't enforce
changes in api package.